Refactor/#12

+++ src/main/java/com/mmp/beacon/commute/query/response/CommuteStatisticsResponse.java
@@ -8,12 +8,14 @@ public record CommuteStatisticsResponse(
             Long id,
             String userId,
             String name
-    ) {}
+    ) {
+    }
 
     public record CommuteStatistics(
-            int presentDays,
-            int lateDays,
-            int absentDays,
-            int totalDays
-    ) {}
+            long presentDays,
+            long lateDays,
+            long absentDays,
+            long totalDays
+    ) {
+    }
 }
